Fetch bound radius from function, not variable
Previously, the plugin would check if a method called ound_radius existed, and then proceed to call to get the variable. Instead of doing it, the plugin will now check if the method exists and call it.

The method must return a floating point number, which will be converted to a float.